Refactor SIMD code.
